feat: implement true board rotation
This rather large update implements actual board rotation.
Prior commits were simply flipping the board on the horizontal axis.

Currently piece movement is working correctly - though things like
error handling for incorrectly choosing a piece or moving to an invalid
location are not yet rescued.

Began some implementation on working in which squares are threatened
and which are not.