refactor: has_moved to parent class
Several pieces need to be checked to see if they've moved for various
reasons.  Makes more sense to just have has_moved available on the
chess_piece class as an accessible (writable) attribute so that all
pieces inherit it and the ones that utilize it will have access to it.