refactor: separate concerns
piece_handler was doing absolutely too much.  Spent some time
tearing it apart and separating out areas of concerns.  Built
a new pathing module & class to deal specifically with paths and
rehomed several methods into square_handler where it just made more
sense to put things.

There are a lot of little changes in this commit that add up to:
We are where we were before I started trying to add in the threat check
code.